Major refactoring and overall improvement of code structure, along with some bug fixes. More to come.
