--- src/graph_type.py
@@ -1,9 +1,0 @@
-from enum import Enum
-
-class GraphType(Enum):
-    ListSpectralLines = 1
-    AbsorptionCoefficient = 2
-    AbsorptionCrossSection = 3
-    AbsorptionFunction = 4
-    TransmittanceFunction = 5
-    RadianceFunction = 6
