+++ src/windows/graph_display_window.py
@@ -1,0 +1,42 @@
+from PyQt5 import QtCore
+from utils.log import *
+from widgets.graph_display_window_gui import *
+from worker.hapi_worker import *
+
+
+class GraphDisplayWindow(QtCore.QObject):
+    done_signal = QtCore.pyqtSignal(object)
+
+    # Data should never be None / null, since if there is no data selected a new
+    # graphing window shouldn't even be opened.
+    def __init__(self, work_object, parent):
+        super(GraphDisplayWindow, self).__init__()
+        self.parent = parent
+        # graph_thread.start()
+        self.gui = GraphDisplayWindowGui()
+        # graph_thread.join()
+        # if len(graph_thread.errors) == 0:
+        #     self.gui.add_graph(graph_thread.x, graph_thread.y)
+        # else:
+        #     for error in graph_thread.errors:
+        #         err_(str(error))
+        self.worker = HapiWorker(work_object, self.plot)
+        self.worker.start()
+        self.done_signal.connect(lambda: parent.done_graphing())
+
+    def plot(self, data):
+        self.done_signal.emit(0)
+        try:
+            (x, y) = (data['x'], data['y'])
+            self.gui.add_graph(x, y, data['title'], data['titlex'], data['titley'])
+        except Exception as e:
+            err_log(e)
+
+    def close(self):
+        for window in self.child_windows:
+            if window.is_open:
+                window.close()
+        self.gui.close()
+
+    def open(self):
+        self.gui.open()
